+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## CHANGELOG
+### UNRELEASED
+- Calls reader 0.2.4 with dyslexia friendly font
 
 ### v3.0.0